{{cookiecutter.project_slug}}/requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
jinja2>=3.1.0
feedparser>=6.0.0
